requirements.txt
Flask==0.12
smartystreets.py==0.4.0
watson-developer-cloud==0.26.0
